MQTT5: ReceiveMaximum flow control option support draft

+++ System.Net.Mqtt.Server/Protocol/V3/MqttServerSession4.cs
@@ -4,7 +4,7 @@ public sealed class MqttServerSession4 : MqttServerSession3
 {
     public MqttServerSession4(string clientId, NetworkTransportPipe transport,
         ISessionStateRepository<MqttServerSessionState4> stateRepository,
-        ILogger logger, int maxUnflushedBytes, int maxInFlight) :
+        ILogger logger, int maxUnflushedBytes, ushort maxInFlight) :
         base(clientId, transport, stateRepository, logger, maxUnflushedBytes, maxInFlight)
     { }
 }
